+++ ChangeLog.md
@@ -1,5 +1,15 @@
 # OldAS ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- Added `SubscribeResult.succeeded`.
+  ([#56](https://github.com/davep/oldas/pull/56))
+- Added `SubscribeResult.__bool__` as an alias for
+  `SubscribeResult.succeeded`.
+  ([#56](https://github.com/davep/oldas/pull/56))
+
 ## v0.12.0
 
 **Released: 2026-02-08**